requirement.txt
altair==5.1.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.7.0
branca==0.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
folium==0.15.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.48.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mojimoji==0.0.12
numpy==1.24.4
packaging==21.3
pandas==2.0.3
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
plotly==5.18.0
proto-plus==1.22.3
protobuf==3.20.3
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.28.2
streamlit-folium==0.15.1
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
zipp==3.17.0
